% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.in.wages.R
\name{change.in.wages}
\alias{change.in.wages}
\title{Change in Wages}
\usage{
change.in.wages(x, y)
}
\arguments{
\item{x}{the initial wage or NA}

\item{y}{the ending wage or NA}
}
\value{
This function returns y-x
}
\description{
This function takes in two numbers x and y which represent wages and determines
the difference of y - x. If x or y is NA then it is changed to zero.
}
