% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.to.logical.R
\name{change.to.logical}
\alias{change.to.logical}
\title{Change to Logical}
\usage{
change.to.logical(df, col, from_list = F, is_column = F)
}
\arguments{
\item{df}{Dataframe that contains the column to change to logical.}

\item{col}{Column in df to change to logical}

\item{from_list}{specied if using map funciton and going through a list of strings in which each string has a column name.}

\item{is_colum}{specified is the return value is a column. Use if implementing map function.}
}
\value{
A df with a column/vector with R logical representations.
}
\description{
This function takes a column and changes a true or false representation to the R Language logical(boolean) native representation.
}
\examples{
column1 <- c(0, 1, 1, 1, 0)
column2 <- c("F", "F", "T", "F")
column3 <- c("no", "no", "no", "yes", "no")

df$column1 <- column1
df$column1 <- column1
df$column1 <- column1

column1 <- change.to.logical(df, column1)
column2 <- change.to.logical(df, column2)
column3 <- change to.logical(df, column3)

}
