% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.to.na.R
\name{change.to.na}
\alias{change.to.na}
\title{Change to NA}
\usage{
change.to.na(df, column, change = 0)
}
\arguments{
\item{column}{column name to change values to NAs}

\item{change}{value to change to NA (default is 0)}
}
\value{
the return value is the user-defined value
}
\description{
This changes values to NA; the default is to change 0's to NAs
}
\examples{
v1 <- c(0, 1, 0)
v1 <- as.factor(v1)
v2 <- change.to.na(v1, change = 0)
df$col <- change.to.na(df$col, change = -999)
}
