% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.to.vect.str.R
\name{change.to.vect.str}
\alias{change.to.vect.str}
\title{Change to vector string.}
\usage{
change.to.vect.str(df, vector)
}
\arguments{
\item{df}{A dataframe with the columns we want to put in a vector string.}

\item{vector}{Numeric vector with the number of columns from the dataframe we want to make a vector string.}
}
\value{
A vector of strings with the names of the columns from df.
}
\description{
Returns a vector of strings(character vector) with the name of the columns in vector.
}
\examples{
df <- my_dataframe

string.vectors <- change.to.vect.str(df, c(1:20))

}
