% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.func.R
\name{clear.func}
\alias{clear.func}
\title{Clear}
\usage{
clear.func(.e)
}
\arguments{
\item{e}{Defaults to enviroment}
}
\description{
Clears all the functions from the enviroment.
}
\examples{
#Delete all functions from the enviroment.
clear.func()


}
