% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.scars.tats.R
\name{count.scars.tats}
\alias{count.scars.tats}
\title{DPS function to count scars and tattoos.}
\usage{
count.scars.tats(.df)
}
\arguments{
\item{df}{Should be dps1.}
}
\value{
dps1 is updated with tats and scars columns.
}
\description{
This function is used in dps1 prep to create column scars and tattoos.
}
\examples{
#In prep file
count.scars.tats(dps1)

}
