% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.to.var.R
\name{count.to.var}
\alias{count.to.var}
\title{count.to.var}
\usage{
count.to.var(df, col, nested.factor, var.names.list, factor.levels.list,
  na2 = 0)
}
\arguments{
\item{df}{dataframe name}

\item{col}{String name of nested column. (With quotes."")}

\item{nested.factor}{name of nested factor within the nested column}

\item{var.names.list}{vector of new variable names}

\item{factor.levels.list}{vector of factor levels to extract}

\item{na2}{replace na with; default is 0}
}
\value{
A dataframe with new variables with counts of those
}
\description{
This function takes a nested factor column and creates new (unnested) variables with counts of the factor levels
}
\examples{
tmp2 <- count.to.var(df = tmp,
                  col = "pre",
                  nested.factor = "level",
                   var.names.list = var.names.list,
                   factor.levels.list = factor.levels.list,
                   na2 = 0)
}
