% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.plot.R
\name{density.plot}
\alias{density.plot}
\title{Density Plot}
\usage{
\method{density}{plot}(dat, cont.out, cat.pre, alpha = 0.5, filename,
  plot.percentage = 100)
}
\arguments{
\item{dat}{Dataframe to take info from.}

\item{cont.out}{List of continuous outcomes}

\item{cat.pre}{List of categorical Predictors}

\item{filename}{Name of the file}
}
\value{
Saves density plots on a pdf with the given filename
}
\description{
This function takes a list of categorical predictors and a list of continuous outcomes to produce density plots.
}
\examples{
density.plot(data=df, cont.out = continous.outcomes, cat.pre = categorical.predictors, alpha=1, filename="Density Plot")

}
