% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.by.date.R
\name{filter.by.date}
\alias{filter.by.date}
\title{Filter by Date.}
\usage{
filter.by.date(listname, task, col.name, new.col.name = NA, date1 = NA,
  date2 = NA)
}
\description{
See vignette: vignette("filter.by.date","wangr")
}
