% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.min.max.R
\name{fix.min.max}
\alias{fix.min.max}
\title{Fix min max}
\usage{
fix.min.max(.column, .min, .max, change.min.to, change.max.to)
}
\arguments{
\item{column}{A list where each element is of an arbitrary type called A of is NA}

\item{min}{A value of the same arbitrary type A as in column where all values in column that are less than it, will be changed.
NOTE: If min is equal to NA then no element in column will be considered to be less than min.
Default Value: NA}

\item{max}{A value of the same arbitrary type A as in column where all values in column that
are greater than it will be changed.
NOTE: If max is equal to NA then no element in column will be considered to be greater than max.
Default Value: NA}

\item{change.min.to}{A value of the same arbitrary type A as in column where all values in column that are less than min will be changed to the value stored in change.min.to.
Default Value: NA}

\item{change.max.to}{A value of the same arbitrary type A as in column where all values in
column that are greater than max will be changed to the value stored in change.max.to.
Default Value: NA}
}
\value{
Column with fixed min and max.
}
\description{
Changes all values in column that are less than min to the value in
}
\examples{
See vignette for example.

}
