% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.id.pswd.R
\name{gen.id.pswd}
\alias{gen.id.pswd}
\title{Generate Id-Passwords}
\usage{
gen.id.pswd(df, str.len = 10)
}
\arguments{
\item{df}{A dataframe to put the two new columns}

\item{str.len}{The length of each password string in string.id column. Defaults to 10.}
}
\value{
A dataframe with two new columns, number.id and string.id.
}
\description{
This function appends two columns to a given dataframe.
One column will be of random numbers and the other column of random strings that include numbers and letters.
}
\examples{
df <- my.dataframe

df <- get.id.pswd(df)
df <- get.id.pswd(df=df, str.len=7)


}
