% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.count.R
\name{get.count}
\alias{get.count}
\title{Get Count}
\usage{
get.count(data, column, nested.column)
}
\arguments{
\item{data}{the data that contains the column to be counted}

\item{column}{the column that is nested}

\item{nested.column}{the column within column that needs to be counted}
}
\value{
a tibble that contains one column called entries which has the thing being counted and
count which contains the number of times that the value of entry was found. Note that even if the
same value was found multiple times in nested.column for one row in column, it will only count once.
}
\description{
get.count counts the number of rows with a given entry in the nested column.
}
