% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.wage.data.R
\name{get.wage.data}
\alias{get.wage.data}
\title{Get Wage Data}
\usage{
get.wage.data(jobs, incar.end.date)
}
\arguments{
\item{jobs}{a tibble containing all the jobs for a person containing the following columns: "start.date" - the date the jobs started, "dot.code" - the dot code for that job.}

\item{incar.end.date}{a date object that is the day the person was released from prison.}
}
\value{
The function returns a tibble where the first job the person had is the first row and the job the person had at
one year after first being employed is the second row. If the person was never employed the function returns
NULL.
}
\description{
The function takes in all the jobs for a person as a tibble and the incarceration end date for the person.
Then it returns a tibble where the first job the person had is the first row and the job the person had at
one year after first being employed is the second row. If the person was never employed the function returns
NULL.
}
