% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.zeros.R
\name{get.zeros}
\alias{get.zeros}
\title{Get Zeros}
\usage{
get.zeros(num.zeros)
}
\arguments{
\item{num.zeros}{the number of zeros the resulting string should have}
}
\value{
A string of num.zeros number of zeros.
}
\description{
This function takes in a number and returns a string composed of that many zeros.
}
