% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.desc.tbl.R
\name{make.desc.tbl}
\alias{make.desc.tbl}
\title{Descriptive Table}
\usage{
make.desc.tbl(df, col.names, factors.vector = NULL, numerics.vector = NULL,
  name = "descriptive.table")
}
\arguments{
\item{df}{A dataframe to base the column names to compare.}

\item{col.names}{Names of the columns from the final table.}

\item{name}{Name of the resulting dataframe.}

\item{factors.vectors}{Column names from the dataframe that are factors. Will be rows in resulting dataframe. (Defaults to NULL)}

\item{numerics.vectors}{Column names from the dataframe that are numerics. Will be rows in resulting dataframe. (Defaults to NULL)}
}
\description{
This will generate a table that describes columns in a dataframe.
change.to.vect.str is a useful function to use with make.desc.tbl
}
\examples{
#normal
make.desc.tbl(df, col.names = col.names,
             factors.vector = factors.vector,
             numerics.vector = numerics.vector)

factor null
make.desc.tbl(df, col.names = col.names,
             factors.vector = NULL,
             numerics.vector = numerics.vector,
             name = "no.factors")

numerics null
make.desc.tbl(df,
             col.names = col.names,
             factors.vector = factors.vector,
             numerics.vector = NULL,
             name = "no.numerics")
}
