% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.R
\name{mv}
\alias{mv}
\title{Change name of objects.}
\usage{
mv(old.name, new.name)
}
\arguments{
\item{old.name}{The name of the object to change name.}

\item{new.name}{The new name for the object.}
}
\value{
Name of object is changed in the enviroment.
}
\description{
This function changes the name of an object without having to create a new one.
}
\examples{
a <- data.frame()

#Will change the name from a to c.
mv(a,c)

}
