% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci.R, R/rci_copy.r
\name{rci}
\alias{rci}
\alias{rci}
\title{RCI}
\usage{
rci(data, pretest, posttest, reliability, increase = FALSE, name,
  graph.title = NA, plot.percentage = 100, categories = 7)

rci(data, pretest, posttest, reliability, increase = FALSE, name,
  graph.title = NA, plot.percentage = 100, categories = 7)
}
\arguments{
\item{data}{the data set}

\item{pretest}{the column name for pretest}

\item{posttest}{the column name for posttest}

\item{reliability}{the reliablity score (the number should be less than 1)}

\item{increase}{True if the increase is good, otherwise False meaning degrease is good}

\item{name}{a string that will be used to create column names to specify which test the rci/difference result and test differences are for}

\item{graph.title}{a string that is the title that will appear on the graph}

\item{plot.percentage}{a number (1-100) of the percentege of point to be plotted}

\item{categories}{number of categories to be provided in the key. Should only be 3 or 7.}

\item{data}{the data set}

\item{pretest}{the column name for pretest}

\item{posttest}{the column name for posttest}

\item{reliability}{the reliablity score (the number should be less than 1)}

\item{increase}{True if the increase is good, otherwise False meaning degrease is good}

\item{name}{a string that will be used to create column names to specify which test the rci/difference result and test differences are for}

\item{graph.title}{a string that is the title that will appear on the graph}

\item{plot.percentage}{a number (1-100) of the percentege of point to be plotted}

\item{categories}{number of categories to be provided in the key. Should only be 3 or 7.}
}
\value{
the results of the function as well as a graph that is created

the results of the function as well as a graph that is created
}
\description{
This function calculates RCI and creates a graph of the results.

This function calculates RCI and creates a graph of the results.
}
