% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.nested.col.R
\name{remove.nested.col}
\alias{remove.nested.col}
\title{Remove Nested Column}
\usage{
remove.nested.col(df, nested.col, col.to.rm)
}
\arguments{
\item{nested.col}{the column that is nested}

\item{col.to.rm}{the column within the nested column to be removed. (Name of column with quotes.)}
}
\value{
the nested column without the column that was supposed to be removed
}
\description{
This funciton removes a column from a tibble which is within a nested column of another tibble.
}
\examples{
data <- a data frame with a nested column called nest, the nested column has two columns, a and b.

result <- remove.nested.col(data$nest, "a")

result will the column data$nest without the column a.

}
