% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtable.R
\name{rtable}
\alias{rtable}
\title{Triangle of Correlation Matrix}
\usage{
rtable(.x, method, removeTriangle, result, stars)
}
\arguments{
\item{x}{Correlation matrix.}

\item{method}{Correlation method. "pearson"" or "spearman"" is supported. Defaults to "pearson".}

\item{removeTriangle}{remove "upper" or "lower" triangle. Defaults to "upper"}

\item{stars}{True if want to see stars, else False. Defaults to False.}

\item{results}{"html" or "latex". Defaults to html}
}
\value{
Table displayed in html or latex format.
}
\description{
Creates a specified triangle (upper or lower) of a correlation matrix and saves it into html or latez format
with or without "stars" (asterisks).

modified from
http://www.sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package
}
\examples{
#Pearson, remove upper triangle, html, no stars
rtable(correlationMatrix, "pearson", "upper","html", False)

#Spearman, remove lower triangle, latex, with stars
rtable(correlationMatrix, "spearman", "lower", "latex", True)

}
