% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.plot.R
\name{scatter.plot}
\alias{scatter.plot}
\title{Scatter Plot}
\usage{
scatter.plot(dat, cat.pre, cont.pre, cont.out, alpha = 0.5, filename,
  plot.percentage = 100)
}
\arguments{
\item{dat}{Dataframe to take info from.}

\item{cat.pre}{List of categorical Predictors}

\item{cont.out}{List of continuous outcomes}

\item{filename}{Name of the file}
}
\value{
Saves scatter plots on a pdf with the given filename
}
\description{
This function takes a list of categorical predictors and a list of continuous outcomes to produce scatter plots.
}
\examples{
scatter.plot(data=df, cont.out = continous.outcomes, cat.pre = categorical.predictors, alpha=1, filename="Density Plot")

}
