% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss2date.R
\name{spss2date}
\alias{spss2date}
\title{Spss to date}
\usage{
spss2date(.col)
}
\arguments{
\item{col}{Dataframe to change the date format from spss.}
}
\value{
A column with dates in R format.
}
\description{
This function must be used if the date values are stored are SPSS "dates",
}
\examples{
#Let dates.in.spss be a column with spss dates
dates <- spss2date(dates.in.spss)


}
