% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabe.grade.lvl.R
\name{tabe.grade.lvl}
\alias{tabe.grade.lvl}
\title{TABE Grade Lvl}
\usage{
tabe.grade.lvl(df, col.names)
}
\arguments{
\item{df}{data frame}

\item{col.names}{column names of TABE grade equivalents}
}
\value{
the return is one new column for every column listed in col.names, with .lvl added to the original col.name
}
\description{
This aggregates TABE grade equivalents to grade levels
}
\examples{

}
