% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.single.col.R
\name{unnest.single.col}
\alias{unnest.single.col}
\title{Unnest Single Column}
\usage{
unnest.single.col(column)
}
\arguments{
\item{column}{a single column where each element contains either Null,
a tibble with a single column and single row, or a tibble with zero rows}
}
\value{
a single list where both Nulls and tibbles with zero rows return NA otherwise
the value is the single value in the tibble
}
\description{
unnest.single.col takes a single column where each element contains either Null,
a tibble with a single column and single row, or a tibble with zero rows and then
it returns a single list where both Nulls and tibbles with zero rows return NA otherwise
the value is the single value in the tibble.
}
