% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang.naming.R
\name{wang.naming}
\alias{wang.naming}
\title{Wang Naming}
\usage{
wang.naming(.df)
}
\arguments{
\item{df}{Dataframe to change the column names}
}
\value{
Dataframe with names changed to Dr. Wang's naming convention.
}
\description{
Change names in dataframe, to Dr. Wang's naming convention.
}
\examples{
#Returns dataframe with names changed.
df <- wang.naming(df)

}
